Edges vitis code fix

+++ Edges/Edges/hls/impl/ip/drivers/vitis_convolution_v1_0/src/xvitis_convolution_linux.c
@@ -18,7 +18,7 @@
 
 /**************************** Type Definitions ******************************/
 typedef struct {
-    u64 addr;
+    u32 addr;
     u32 size;
 } XVitis_convolution_uio_map;
 
@@ -126,11 +126,11 @@ int XVitis_convolution_Initialize(XVitis_convolution *InstancePtr, const char* I
     }
 
     // NOTE: slave interface 'Control' should be mapped to uioX/map0
-    InstancePtr->Control_BaseAddress = (u64)mmap(NULL, InfoPtr->maps[0].size, PROT_READ|PROT_WRITE, MAP_SHARED, InfoPtr->uio_fd, 0 * getpagesize());
+    InstancePtr->Control_BaseAddress = (u32)mmap(NULL, InfoPtr->maps[0].size, PROT_READ|PROT_WRITE, MAP_SHARED, InfoPtr->uio_fd, 0 * getpagesize());
     assert(InstancePtr->Control_BaseAddress);
 
     // NOTE: slave interface 'Control_r' should be mapped to uioX/map1
-    InstancePtr->Control_r_BaseAddress = (u64)mmap(NULL, InfoPtr->maps[1].size, PROT_READ|PROT_WRITE, MAP_SHARED, InfoPtr->uio_fd, 1 * getpagesize());
+    InstancePtr->Control_r_BaseAddress = (u32)mmap(NULL, InfoPtr->maps[1].size, PROT_READ|PROT_WRITE, MAP_SHARED, InfoPtr->uio_fd, 1 * getpagesize());
     assert(InstancePtr->Control_r_BaseAddress);
 
     InstancePtr->IsReady = XIL_COMPONENT_IS_READY;
