+++ Edges/Edges/hls/impl/misc/drivers/vitis_convolution_v1_0/src/xvitis_convolution.h
@@ -45,14 +45,14 @@ typedef struct {
 #else
     u16 DeviceId;
 #endif
-    u64 Control_BaseAddress;
-    u64 Control_r_BaseAddress;
+    u32 Control_BaseAddress;
+    u32 Control_r_BaseAddress;
 } XVitis_convolution_Config;
 #endif
 
 typedef struct {
-    u64 Control_BaseAddress;
-    u64 Control_r_BaseAddress;
+    u32 Control_BaseAddress;
+    u32 Control_r_BaseAddress;
     u32 IsReady;
 } XVitis_convolution;
 
@@ -101,10 +101,10 @@ u32 XVitis_convolution_IsReady(XVitis_convolution *InstancePtr);
 void XVitis_convolution_EnableAutoRestart(XVitis_convolution *InstancePtr);
 void XVitis_convolution_DisableAutoRestart(XVitis_convolution *InstancePtr);
 
-void XVitis_convolution_Set_input_img(XVitis_convolution *InstancePtr, u64 Data);
-u64 XVitis_convolution_Get_input_img(XVitis_convolution *InstancePtr);
-void XVitis_convolution_Set_output_img(XVitis_convolution *InstancePtr, u64 Data);
-u64 XVitis_convolution_Get_output_img(XVitis_convolution *InstancePtr);
+void XVitis_convolution_Set_input_img(XVitis_convolution *InstancePtr, u32 Data);
+u32 XVitis_convolution_Get_input_img(XVitis_convolution *InstancePtr);
+void XVitis_convolution_Set_output_img(XVitis_convolution *InstancePtr, u32 Data);
+u32 XVitis_convolution_Get_output_img(XVitis_convolution *InstancePtr);
 u32 XVitis_convolution_Get_kernel_BaseAddress(XVitis_convolution *InstancePtr);
 u32 XVitis_convolution_Get_kernel_HighAddress(XVitis_convolution *InstancePtr);
 u32 XVitis_convolution_Get_kernel_TotalBytes(XVitis_convolution *InstancePtr);
